Move HomePage to self file
